backend/pyproject.toml
[tool.poetry]
name = "climbingcrux-model"
version = "0.1.0"
description = ""
authors = ["Michał Kurcewicz <mkurc1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
jupyterlab = "^4.1.2"
jupyter = "^1.0.0"
ultralytics = "^8.2.25"
opencv-python = "^4.9.0.80"
opencv-contrib-python = "^4.9.0.80"
scipy = "^1.12.0"
imutils = "^0.5.4"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
torch = "2.2.2"
python-multipart = "^0.0.9"
google-auth = "^2.29.0"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
